refactor(frontend): consolidate readonly SQL code display
Extract getReadonlyCodeDisplay so vertical layout, snippets panel, and
slides read-only views share one SQL unwrap/highlight path instead of
duplicating SQLParser logic.
