+++ frontend/src/components/editor/renderers/vertical-layout/vertical-layout.tsx
@@ -10,7 +10,6 @@ import {
   Loader2Icon,
   MoreHorizontalIcon,
 } from "lucide-react";
-import { SQLParser } from "@marimo-team/smart-cells";
 import type React from "react";
 import { memo, useRef, useState } from "react";
 import { z } from "zod";
@@ -30,6 +29,7 @@ import { outputIsLoading, outputIsStale } from "@/core/cells/cell";
 import type { CellId } from "@/core/cells/ids";
 import { isOutputEmpty } from "@/core/cells/outputs";
 import type { CellData, CellRuntimeState } from "@/core/cells/types";
+import { getReadonlyCodeDisplay } from "@/core/cells/readonly-code-display";
 import { MarkdownLanguageAdapter } from "@/core/codemirror/language/languages/markdown";
 import { useResolvedMarimoConfig } from "@/core/config/config";
 import { CSSClasses, KnownQueryParams } from "@/core/constants";
@@ -386,7 +386,7 @@ const VerticalCell = memo(
       // Hide the code if it's pure markdown and there's an output, or if the code is empty
       const hideCode = shouldHideCode(code, output);
       // Only unwrap SQL when the code will actually be rendered.
-      const display = hideCode ? null : readonlyDisplay(code);
+      const display = hideCode ? null : getReadonlyCodeDisplay(code);
 
       return (
         <div
@@ -499,21 +499,4 @@ export function shouldHideCode(code: string, output: OutputMessage | null) {
   const isPureMarkdown = new MarkdownLanguageAdapter().isSupported(code);
   const hasOutput = output !== null && !isOutputEmpty(output);
   return (isPureMarkdown && hasOutput) || code.trim() === "";
-}
-
-const sqlParser = new SQLParser();
-
-/**
- * Unwrap SQL cells to their inner query so read-only views highlight them as
- * SQL instead of showing the raw `mo.sql(...)` wrapper.
- */
-export function readonlyDisplay(code: string): {
-  code: string;
-  language: "python" | "sql";
-} {
-  const trimmed = code.trim();
-  if (sqlParser.isSupported(trimmed)) {
-    return { code: sqlParser.transformIn(trimmed).code, language: "sql" };
-  }
-  return { code, language: "python" };
 }
